backend/point_of_sale/requirements/base.txt
Django==3.0.2
djangorestframework==3.11.0
psycopg2-binary==2.8.4
python-decouple==3.3
dj_database_url==0.5.0
django-cors-headers==3.2.1
